Revert "code seperation files"
