--- include/blockScoring.hpp
@@ -1,3 +1,0 @@
-#pragma once
-
-void scoreBlocks(int numberOfBlocks, bool colorsort);
