--- include/colorSort.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-extern int BLUE_MAX;
-extern int RED_MAX;
-extern int BLUE_MIN;
-extern int RED_MIN;
-extern int OPP_MIN;
-extern int OPP_MAX;
-
-void colorSort();
