--- include/distanceSensor.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-
-float wallDistance(bool shouldPrint, bool useRightSensor);
