--- include/midGoalScoring.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-
-void scoreMidGoal();
