--- include/oneGoal.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-
-void one_goal_left();
-void one_goal_right();
