--- include/pidTesting.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-
-void turnTesting(bool isCW);
-void driveTesting(bool isForward);
