+++ include/setup.hpp
@@ -11,8 +11,8 @@
 inline pros::Controller controller(pros::E_CONTROLLER_MASTER); //controller
 
 // drivetrain
-inline pros::MotorGroup right_dt({7,-8 , 9}, pros::MotorGearset::blue);    // right side
-inline pros::MotorGroup left_dt({-14,-16,15}, pros::MotorGearset::blue);  // left side
+inline pros::MotorGroup left_dt({-7, -6, 5}, pros::MotorGearset::blue);    // left side
+inline pros::MotorGroup right_dt({14, 15, -16}, pros::MotorGearset::blue);  // right side
 
 // drivetrain config for movement functions
 // (wheel size, width, rpm, etc.)
@@ -24,19 +24,18 @@ inline lemlib::Drivetrain drivetrain(&left_dt,
 							  2
 );
 
-inline pros::Imu imu(1);
+inline pros::Imu imu(20);
 // declare sensors needed for odom setup
 /*inline pros::Rotation horizontalEnc(17);
 inline pros::Rotation verticalEnc(18);
-*/
 
-inline pros::Rotation verticalEnc(-6);
-//configure tracking wheels
-//inline lemlib::TrackingWheel horizontalTrackingWheel(&horizontalEnc, lemlib::Omniwheel::NEW_2 * 24/25.2, -5);
-inline lemlib::TrackingWheel verticalTrackingWheel(&verticalEnc, lemlib::Omniwheel::NEW_275, 0);
-//inline lemlib::TrackingWheel verticalTrackingWheel(&verticalEnc, lemlib::Omniwheel::NEW_2 * 24/25.2, 1.5);
+
+// configure tracking wheels
+inline lemlib::TrackingWheel horizontalTrackingWheel(&horizontalEnc, lemlib::Omniwheel::NEW_2 * 24/25.2, -5);
+inline lemlib::TrackingWheel verticalTrackingWheel(&verticalEnc, lemlib::Omniwheel::NEW_2 * 24/25.2, -0.475);*/
+
 // configure entire odom setup
-inline lemlib::OdomSensors sensors(&verticalTrackingWheel,
+inline lemlib::OdomSensors sensors(nullptr,
 							nullptr,
 							nullptr,
 							nullptr,
@@ -44,9 +43,9 @@ inline lemlib::OdomSensors sensors(&verticalTrackingWheel,
 );
 
 // lateral pid
-inline lemlib::ControllerSettings lateral_controller(6.25, //proportional gain (kP) //5.5
-                                            0, // integral gain (kI) //0.5
-                                        25, // derivative gain (kD) //18
+inline lemlib::ControllerSettings lateral_controller(5.25, //proportional gain (kP) //5.5
+                                            0.55, // integral gain (kI) //0.5
+                                        21, // derivative gain (kD) //18
                                          1,//windup
                                            0.5, // small error range, in inches
                                           100, // small error range timeout, in milliseconds
@@ -56,9 +55,9 @@ inline lemlib::ControllerSettings lateral_controller(6.25, //proportional gain (
 );
 
 // angular pid
-inline lemlib::ControllerSettings angular_controller(5, // proportional gain (kP)
-                                              0, // integral gain (kI)
-                                              43.5, // derivative gain (kD)
+inline lemlib::ControllerSettings angular_controller(3.5, // proportional gain (kP) was 3.25 (reverted to 3.5)
+                                              0.4, // integral gain (kI)
+                                              24, // derivative gain (kD) was 22 
                                               3.5, // anti windup
                                               0.5, // small error range, in degrees
                                                 50, // small error range timeout, in milliseconds
@@ -75,12 +74,12 @@ inline lemlib::ControllerSettings angular_controller(5, // proportional gain (kP
 inline lemlib::ExpoDriveCurve
     throttle_curve(3,    // joystick deadband out of 127
                    0,   // minimum output where drivetrain will move out of 127
-                   1.017 // expo curve gain
+                   1.02 // expo curve gain
     );
 
 // input curve for steer input during driver control
 inline lemlib::ExpoDriveCurve
-    steer_curve(60,    // joystick deadband out of 127
+    steer_curve(10,    // joystick deadband out of 127
                 15,   // minimum output where drivetrain will move out of 127
                 1.016 // expo curve gain
     );
@@ -96,22 +95,20 @@ inline lemlib::Chassis chassis(drivetrain, // drivetrain settings
 
 // declare additional motors, sensors, and pnuematics here
 
-//negative means regular direction is flipped
 //intake groups
-inline pros::Motor frontIntake(5, pros::MotorGearset::blue);
+inline pros::Motor frontIntake(2, pros::MotorGearset::blue);
 inline pros::Motor middleRollers(-10);
 inline pros::Motor colorSortRoller(4);
-inline pros::Motor scoringRoller(-11);
+inline pros::Motor scoringRoller(-19);
 
 //extensions
-// false means open and true means close
-inline pros::adi::DigitalOut scoringGate('A');
+inline pros::adi::DigitalOut scoringBar('A');
 inline pros::adi::DigitalOut loaderMech('B');
-inline pros::adi::DigitalOut wingMech('C');
+inline pros::adi::DigitalOut wingMech('H');
 inline pros::adi::DigitalOut parkMech('D');
 
 // clog sensors
-inline pros::Optical topOptical(4);
+inline pros::Optical topOptical(18);
 inline pros::Distance bottomDistance(8);
 inline pros::Distance leftDistance(17);
-inline pros::Distance rightDistance(13);
+inline pros::Distance rightDistance(9);
