--- include/twoGoal.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-
-void two_goal_RIGHT();
-void two_goal_LEFT();
