+++ project.pros
@@ -1,7 +1,7 @@
 {
   "py/object": "pros.conductor.project.Project",
   "py/state": {
-    "project_name": "newbot",
+    "project_name": "12/6 comp ",
     "target": "v5",
     "templates": {
       "LemLib": {
@@ -324,7 +324,7 @@
     },
     "upload_options": {
       "slot": 1,
-      "description": "new bot code",
+      "description": "comp. intake lock in",
       "icon": "X"
     },
     "use_early_access": false
