+++ src/op_control.cpp
@@ -3,15 +3,15 @@
 #include "pros/llemu.hpp"
 #include "pros/misc.h"
 #include "setup.hpp"
-#include "op_control.hpp"
+
 #include <fstream>
 #include <string>
 
 //STATES
 //extensions
-bool openGate = true;
+bool isScoringBarUp = false;
 bool isLoaderExtended = false;
-bool isWingsOut = true;
+bool isWingsOut = false;
 //core func
 bool isIntakeOn = false;
 bool shouldSC = false;
@@ -52,7 +52,7 @@ void handleIOCommands() {
     isIntakeOn = !isIntakeOn;
   }
 
-  if (controller.get_digital(pros::E_CONTROLLER_DIGITAL_A)) { // when L2 is held the system reverses when let go it returns to the state of L1 toggle
+  if (controller.get_digital(pros::E_CONTROLLER_DIGITAL_L2)) { // when L2 is held the system reverses when let go it returns to the state of L1 toggle
     frontIntake.move(-127);
     middleRollers.move(-127);
     scoringRoller.move(-127);
@@ -66,18 +66,18 @@ void handleIOCommands() {
 
   // hold behavior while R1 is held
   if (controller.get_digital(pros::E_CONTROLLER_DIGITAL_R1)) { // when R1 is held the system runs forward with scoring bar deployed when let go returns to the state of L1 toggle
-    scoringGate.set_value(false);
+    scoringBar.set_value(true);
     frontIntake.move(127);
     middleRollers.move(127);
     scoringRoller.move(127);
   } else { // what happens when the R1 is let go off
-    scoringGate.set_value(true);
+    scoringBar.set_value(false);
     
   }
   bool r1_active = controller.get_digital(pros::E_CONTROLLER_DIGITAL_R1);
 
   if (controller.get_digital(pros::E_CONTROLLER_DIGITAL_R2)) { // when R2 is held the system runs forward with scoring roller reversed for mid scoring and when let go returns to the state of L1 toggle
-    scoringGate.set_value(true);
+    scoringBar.set_value(false);
     frontIntake.move(127);
     middleRollers.move(100);
     scoringRoller.move(-50);
@@ -113,7 +113,7 @@ void telemetry() {
         << chassis.getPose().x << ","
         << chassis.getPose().y << ","
         << chassis.getPose().theta << ","
-        << openGate << ","
+        << isScoringBarUp << ","
         << isLoaderExtended << ","
         << isWingsOut << "\n";
   file.close();
@@ -131,7 +131,7 @@ void handleLoaderMechCommands() { //toggle button for loader mech
 }
 
 void handleWingMechCommands() { //toggle button wing mech
-  if (controller.get_digital_new_press(pros::E_CONTROLLER_DIGITAL_L2)) { //if the controller recognizes a new press from the B button
+  if (controller.get_digital_new_press(pros::E_CONTROLLER_DIGITAL_A)) { //if the controller recognizes a new press from the B button
     isWingsOut = !isWingsOut; //flips the condition of the current state of the wings
     wingMech.set_value(isWingsOut); //sets the physical state to the bool condition of the wings
     if (isWingsOut == false){
