--- src/testRoute.cpp
@@ -1,297 +1,0 @@
-#include "lemlib/chassis/chassis.hpp"
-#include "main.h"
-#include "pros/distance.hpp"
-#include "pros/llemu.hpp"
-#include "setup.hpp"
-#include "testRoute.hpp"
-
-void oldAutonSkills() { 
-    //SETUP
-    chassis.setPose(15,-48,90);
-    frontIntake.move(127);
-    middleRollers.move(127);
-    scoringRoller.move(127);
-    left_dt.set_brake_mode(pros::MotorBrake::brake);
-    right_dt.set_brake_mode(pros::MotorBrake::brake);
-    //Q1
-    chassis.moveToPoint(48,-48,1400,{.minSpeed=60,.earlyExitRange=8});
-    chassis.turnToHeading(180,2000,{.maxSpeed=50},false);
-    loaderMech.set_value(true);
-    pros::delay(100);
-    chassis.moveToPoint(47.5,-65,2500,{.maxSpeed=60},false);
-    chassis.turnToHeading(180,500,{.maxSpeed=80});
-    loaderMech.set_value(false);
-    chassis.moveToPoint(48,-22,2000,{.forwards=false,.minSpeed=80});
-    pros::delay(900);
-    scoringGate.set_value(true);
-    pros::delay(800);
-    frontIntake.move(0);
-    middleRollers.move(0);
-    scoringRoller.move(0);
-    pros::delay(200);
-    scoringGate.set_value(false);
-    chassis.setPose(wallDistance(false,false), -30, chassis.getPose().theta);
-
-    chassis.moveToPoint(48,-36,2000,{.maxSpeed=85});
-    pros::delay(200);
-    wingMech.set_value(true);
-    frontIntake.move(127);
-    middleRollers.move(127);
-    scoringRoller.move(127);
-    scoringGate.set_value(false);
-    chassis.turnToPoint(60,-26,1500,{.maxSpeed=60});
-    chassis.moveToPoint(60,-26,1500,{.minSpeed=60,.earlyExitRange=2});
-    chassis.turnToHeading(0,1500);
-    chassis.moveToPoint(60,36,3000,{.minSpeed=60,.earlyExitRange=8});
-    chassis.turnToPoint(45,48,1500,{.minSpeed=60,.earlyExitRange=3});
-    chassis.moveToPose(45,48,0,2000,{.maxSpeed=60});
-    chassis.turnToHeading(0,1500,{.maxSpeed=80});
-    pros::delay(300);
-    loaderMech.set_value(true);
-    pros::delay(300);
-    chassis.moveToPoint(46,70,3000,{.maxSpeed=60});
-    pros::delay(1000);
-    chassis.moveToPoint(46,22,2500,{.forwards=false,.maxSpeed=80});
-    pros::delay(1300);
-    scoringGate.set_value(true);
-    loaderMech.set_value(false);
-    pros::delay(2000);
-    chassis.setPose(chassis.getPose().x, 30, chassis.getPose().theta);
-    chassis.moveToPose(18,63,-90,2500,{.minSpeed=70,.earlyExitRange=5});
-    chassis.turnToHeading(-85,500,{.earlyExitRange=3},false);
-    scoringGate.set_value(false); 
-    right_dt.move(90);
-    left_dt.move(90);
-    loaderMech.set_value(true);
-    pros::delay(1900);
-    loaderMech.set_value(false);
-    left_dt.move(0);
-    right_dt.move(0);
-    chassis.turnToHeading(-105,1500,{.earlyExitRange=5},false);
-    right_dt.move(-35);
-    left_dt.move(-35);
-    pros::delay(500);
-    left_dt.move(0);
-    right_dt.move(0);
-
-    chassis.turnToHeading(-90,1500,{},false);
-    chassis.setPose(-17,wallDistance(false,true), chassis.getPose().theta);
-    pros::delay(200);
-    chassis.moveToPose(-48,48,-135,3000,{.lead=-0.6,.minSpeed=70});
-    chassis.turnToHeading(0,1000,{},false);
-    chassis.setPose(wallDistance(false,false), chassis.getPose().y, chassis.getPose().theta);
-    chassis.moveToPoint(-48,18,3000,{.forwards=false,.maxSpeed=60});
-    pros::delay(500);
-    scoringGate.set_value(true);
-    loaderMech.set_value(true);
-    pros::delay(1200);
-    chassis.setPose(wallDistance(false,false), 30, chassis.getPose().theta);
-    scoringGate.set_value(false);
-    chassis.moveToPoint(-48,70,3000,{.maxSpeed=60});
-    pros::delay(1000);
-    chassis.moveToPoint(-48,22,2500,{.forwards=false,.maxSpeed=90});
-    pros::delay(1300);
-    scoringGate.set_value(true);
-    loaderMech.set_value(false);
-    pros::delay(2000);
-    chassis.setPose(chassis.getPose().x, 30, chassis.getPose().theta);
-    scoringGate.set_value(false); 
-    chassis.moveToPoint(-48,36,1000,{.maxSpeed=85});
-    pros::delay(200);
-    wingMech.set_value(true);
-    frontIntake.move(127);
-    middleRollers.move(127);
-    scoringRoller.move(127);
-    scoringGate.set_value(false);
-    chassis.turnToPoint(-60,26,1500,{.maxSpeed=60});
-    chassis.moveToPoint(-60,26,1500,{.minSpeed=60,.earlyExitRange=2});
-    chassis.turnToHeading(180,1500);
-    chassis.moveToPoint(-60,-36,3000,{.minSpeed=60,.earlyExitRange=8});
-    chassis.turnToPoint(-45,-48,1500,{.minSpeed=60,.earlyExitRange=3});
-    chassis.moveToPose(-45,-48,180,2000,{.maxSpeed=60});
-    chassis.turnToHeading(180,1500,{.maxSpeed=80},false);
-    pros::delay(300);
-    loaderMech.set_value(true);
-    pros::delay(300);
-    chassis.setPose(wallDistance(false, true), chassis.getPose().y, chassis.getPose().theta);
-    //pros::lcd::print(5, "Wall Dist: %.2f", wallDistance(false, true));
-    pros::lcd::print(6, "Chassis Y: %.2f", chassis.getPose().y);
-    pros::lcd::print(7, "Chassis X: %.2f", chassis.getPose().x);
-    chassis.moveToPoint(-46,-70,3000,{.maxSpeed=60});
-    pros::delay(1000);
-    chassis.moveToPoint(-46,-22,2500,{.forwards=false,.minSpeed=80});
-    pros::delay(1300);
-    scoringGate.set_value(true);
-    loaderMech.set_value(false);
-    pros::delay(2000);
-    //chassis.setPose(chassis.getPose().x, -30, chassis.getPose().theta);
-    scoringGate.set_value(false);
-    wingMech.set_value(true);
-    chassis.moveToPose(-18,-63,90,3000,{.lead=0.5,.minSpeed=80,.earlyExitRange=3});
-    chassis.turnToHeading(95,500,{.earlyExitRange=5},false);
-    loaderMech.set_value(true);
-    right_dt.move(127);
-    left_dt.move(127);
-    pros::delay(1200);
-    left_dt.move(-25);
-    right_dt.move(-25);
-    
-    /*
-        frontIntake.move(127);iop[poio]iu
-    middleRollers.move(127);
-    scoringRoller.move(127);
-    left_dt.set_brake_mode(pros::MotorBrake::brake);
-    right_dt.set_brake_mode(pros::MotorBrake::brake);
-    //Q1
-    chassis.moveToPoint(31.5,0,1400,{.maxSpeed=75});
-    chassis.turnToHeading(180,2000,{.maxSpeed=50},false);
-    loaderMech.set_value(true);
-    pros::delay(500);
-    chassis.moveToPoint(32,-18,2500,{.maxSpeed=60},false);
-    chassis.turnToHeading(180,500,{.maxSpeed=80});
-    loaderMech.set_value(false);
-    chassis.moveToPoint(32,30,2000,{.forwards=false,.maxSpeed=80});
-    pros::delay(900);
-    scoringGate.set_value(true);
-    pros::delay(600);
-    frontIntake.move(0);
-    middleRollers.move(0);
-    scoringRoller.move(0);
-    pros::delay(200);
-    scoringGate.set_value(false);
-    */
-    /*
-    //Q2
-    chassis.moveToPoint(35,7,2000,{.maxSpeed=85});
-    pros::delay(200);
-    wingMech.set_value(true);
-    frontIntake.move(127);
-    middleRollers.move(127);
-    scoringRoller.move(127);
-    scoringGate.set_value(false);
-    chassis.turnToPoint(44,30,1500,{.maxSpeed=60});
-    chassis.moveToPoint(44,30,1500,{.maxSpeed=60});
-    chassis.turnToHeading(0,1500);
-    chassis.moveToPoint(43,93,3000,{.maxSpeed=85});
-    chassis.turnToPoint(28.5,102,1500,{.maxSpeed=60});
-    chassis.moveToPoint(28.5,102,2000,{.maxSpeed=60});
-    chassis.turnToHeading(0,1500,{.maxSpeed=80});
-    pros::delay(300);
-    loaderMech.set_value(true);
-    pros::delay(300);
-    chassis.moveToPoint(26.75,123,3000,{.maxSpeed=60});
-    pros::delay(1000);
-    chassis.moveToPoint(28.5,85,2500,{.forwards=false,.maxSpeed=80});
-    pros::delay(1300);
-    scoringGate.set_value(true);
-    pros::delay(2000);
-    loaderMech.set_value(false);
-    chassis.moveToPoint(28.5,96,3000,{.maxSpeed=60},false);
-    scoringGate.set_value(false);
-    chassis.moveToPoint(28.5,85,2500,{.forwards=false,.minSpeed=80});
-    chassis.moveToPoint(28.5,100,3000,{.maxSpeed=60},false);
-    chassis.turnToHeading(-90,1500);
-    chassis.moveToPoint(-74.5,100,5000,{.maxSpeed=60},false);
-    chassis.turnToHeading(0,1500);
-    loaderMech.set_value(true);
-    chassis.moveToPoint(-76,123,2500,{.maxSpeed=60});
-    pros::delay(1000);
-    chassis.moveToPoint(-77,85,2500,{.forwards=false,.maxSpeed=80});
-    pros::delay(900);
-    scoringGate.set_value(true);
-    pros::delay(500);
-    frontIntake.move(0);
-    middleRollers.move(0);
-    scoringRoller.move(0);
-    pros::delay(200);
-    scoringGate.set_value(false);
-    pros::delay(200);
-    loaderMech.set_value(false);
-    chassis.moveToPoint(-77,102,3000,{.maxSpeed=60},false);
-    chassis.turnToPoint(-92,83,1500);
-    chassis.moveToPoint(-92,83,1500,{.maxSpeed=60});
-    chassis.turnToHeading(180,1500);
-    chassis.moveToPoint(-90,15,3000,{.maxSpeed=60},false);
-    chassis.turnToPoint(-76,0,1500);
-    chassis.moveToPoint(-76,0,2500,{.maxSpeed=60});
-    chassis.turnToHeading(180,1500);
-    frontIntake.move(127);
-    middleRollers.move(127);
-    scoringRoller.move(127);
-    loaderMech.set_value(true);
-    chassis.moveToPoint(-74,-20,2500,{.maxSpeed=60},false);
-    pros::delay(1000);
-    chassis.moveToPoint(-75,35,2500,{.forwards=false,.maxSpeed=80});
-    pros::delay(1300);
-    scoringGate.set_value(true);
-    pros::delay(2000);
-    loaderMech.set_value(false);
-    chassis.moveToPoint(-75,3,3000,{.maxSpeed=60},false);
-    scoringGate.set_value(false);
-    chassis.moveToPoint(-75,35,2500,{.forwards=false,.maxSpeed=80});
-    chassis.moveToPoint(-75,20,3000,{.maxSpeed=60},false);
-    chassis.turnToPoint(-27,-10,2500,{.maxSpeed=80});
-    chassis.moveToPoint(-27,-10,2500,{.maxSpeed=80});
-    chassis.turnToHeading(90,1500);
-    loaderMech.set_value(true);
-    chassis.moveToPoint(0,-10,5000,{.maxSpeed=80},false);
-    /*
-    pros::delay(200);
-    scoringGate.set_value(false);
-    middleRollers.move(127);
-    scoringRoller.move(127);
-    pros::delay(10);
-    chassis.moveToPoint(-61,5,4000,{.maxSpeed=70});
-    chassis.turnToHeading(180,500);
-    loaderMech.set_value(true);
-    chassis.moveToPoint(-61,-20, 3000,{.maxSpeed= 90},false);
-    chassis.moveToPoint(-61,20,3000,{.forwards = false, .maxSpeed= 85});
-    pros::delay(650);
-    scoringGate.set_value(true);
-    pros::delay(2000);
-    chassis.moveToPoint(-61,6,1000,{.maxSpeed= 85});
-    chassis.turnToPoint(-43,0,1500,{.minSpeed=80});
-    scoringGate.set_value(false);
-    /*
-    //Q3
-    chassis.moveToPoint(-43,0,1000,{.maxSpeed=70});
-    chassis.turnToHeading(0,1500);
-    loaderMech.set_value(false);
-    chassis.moveToPoint(-43, 75, 3000,{});
-    chassis.turnToPoint(-60,100,1500,{.maxSpeed=80});
-    chassis.moveToPoint(-60,100,1500,{.maxSpeed=70});
-    chassis.turnToHeading(0,1500);
-    loaderMech.set_value(true);
-    chassis.moveToPoint(-60,115, 4000,{.maxSpeed=70});
-    chassis.moveToPoint(-60.5,75,3000,{.forwards=false,.maxSpeed=90});
-    loaderMech.set_value(false);
-    pros::delay(900);
-    scoringRoller.move(90);
-    scoringGate.set_value(true);
-    pros::delay(4000);
-    chassis.moveToPoint(-60.5,90,2000,{.maxSpeed=85});
-    pros::delay(200);
-    scoringGate.set_value(false);
-    middleRollers.move(127);
-    scoringRoller.move(127);
-    pros::delay(10);
-    //Q4
-    chassis.moveToPoint(34, 90,5000,{.maxSpeed=80});
-    chassis.turnToHeading(0,1000);
-    loaderMech.set_value(true);
-    chassis.moveToPoint(34, 120,5000,{});
-    chassis.moveToPoint(34.5, 75,5000,{.forwards=false,.maxSpeed=90});
-    pros::delay(650);
-    scoringGate.set_value(true);
-    pros::delay(3000);
-    //park
-    chassis.moveToPoint(34.5,100,2000,{.maxSpeed=70});
-    loaderMech.set_value(false);
-    chassis.turnToPoint(17.5,100,1000);
-    chassis.moveToPoint(17.5,100,1500);
-    chassis.turnToHeading(180,1000);
-    chassis.moveToPoint(17.5,-15,1500);
-    chassis.turnToHeading(-90,1000);
-    chassis.moveToPoint(-5,-15,1500);
-    */
-}
